letsRideServer/requirements.txt
Django==4.1.1
djangorestframework-3.14.0
django-cors-headers-3.13.0
psycopg2-binary-2.9.4